fix(ui,lezer): address Copilot PR review comments
generate-icons.ts:
- Remove unused _nx/_ny variables from drawStrokedLine (they were computed
  but not needed for the point-to-segment distance calculation)
- Rename width param to size throughout drawStrokedLine so Y-axis clamping
  (maxY) correctly uses size-1 instead of width-1
- Extract PNG encoding helpers (crc32, uint32BE, chunk, encodePng) into
  scripts/png-encode.ts to keep both files under the 150-line guideline

pattern-serialize.ts:
- Update convertRegexToLezer docstring: remove stale @digit/@asciiLetter
  references; note they are invalid in lezer-generator CLI
